Move CAS1 OAsys endpoints into OAsys Controller
